openai-sdk-project/pyproject.toml
[project]
name = "openai-sdk-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.4.1",
    "openai-agents>=0.0.4",
    "python-dotenv>=1.0.1",
]

[project.scripts]
openai-sdk-project = "openai_sdk_project:main"
openai_run="openai_sdk_project.main:async_test_installation"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
